Update main.py
Adding a select to the database before making request to the websote, to check if the ID already exist in the database. If so, it's not necessary to make request to the website because it's just going to waste time.
Fixing issue where the last name would be added if the next id has no user
New fonction that scan the table for courses name, and adding them to the database if they don't already exist.
Also i put the sessions user id everywhere it was useful in the urls so that the program works no matter the logins you put at the begining

The programm can now successfully retrieve the First and lastname from an id, all the courses names with their coeff and get the grades corresponding to a course and an ID. The filling of the grades databases is left to do otherwise the course database is done and the ID -> name database is also done.